fix: cache key producer. better docs. no more tasks/celery.
Remove everything to do with jobs/tasks/Celery. It was mentioned
this is not needed. Now calls to /titles will produce the arguments
as a response, but will also include a random number to proof cache.

Fix the cache key maker to produce a consistent cache key for
query strings regardless of argument order.